move dev mode test to main test file
